+++ src/main/java/chessEngine/chess/piece/constantMovesPiece/knight/BlackKnight.java
@@ -5,7 +5,7 @@ import chessEngine.chess.piece.PieceColor;
 public class BlackKnight extends Knight{
     private final PieceColor pieceColor = PieceColor.BLACK;
 
-    public BlackKnight(int height, int width) {
+    public BlackKnight(byte height, byte width) {
         super(height, width);
     }
 }
