+++ src/main/java/chessEngine/chess/piece/constantMovesPiece/knight/WhiteKnight.java
@@ -5,7 +5,7 @@ import chessEngine.chess.piece.PieceColor;
 public class WhiteKnight extends Knight{
     private final PieceColor pieceColor = PieceColor.WHITE;
 
-    public WhiteKnight(int height, int width) {
+    public WhiteKnight(byte height, byte width) {
         super(height, width);
     }
 }
