+++ src/main/java/chessEngine/chess/piece/infiniteRangePiece/bishop/Bishop.java
@@ -3,14 +3,14 @@ package chessEngine.chess.piece.infiniteRangePiece.bishop;
 import chessEngine.chess.piece.infiniteRangePiece.InfiniteRangePiece;
 
 public abstract class Bishop extends InfiniteRangePiece {
-    private final int [][] directions = {
+    private final byte [][] directions = {
             {1, 1},
             {1, -1},
             {-1, 1},
             {-1, -1}
     };
 
-    public Bishop(int height, int width) {
+    public Bishop(byte height, byte width) {
         super(height, width);
     }
 }
