+++ src/main/java/chessEngine/chess/piece/infiniteRangePiece/rook/Rook.java
@@ -3,14 +3,14 @@ package chessEngine.chess.piece.infiniteRangePiece.rook;
 import chessEngine.chess.piece.infiniteRangePiece.InfiniteRangePiece;
 
 public abstract class Rook extends InfiniteRangePiece {
-    private final int[][] directions = {
+    private final byte[][] directions = {
             {1, 0},
             {0, 1},
             {-1, 0},
             {0, -1}
     };
 
-    public Rook(int height, int width) {
+    public Rook(byte height, byte width) {
         super(height, width);
     }
 }
